Moved response resolving logic into sepated class.
